Expose IO, Update to new input system
I set out to figure out why double clicking was so hard on 144hz monitors.

I traced this back to a github post complaining about a similar issue, then someone noting that the new input system fixed it, so I went ahead and expose all of ImGuiIO and updated the input code

Anyway turns out we just weren't setting io->DeltaTime in the setEvents function. But that's all deprecated now so we can just set it directly now!